builtin-blame.c: allow more than 16 parents
This removes the hardcoded 16 parents limit from git-blame by allowing the
parent array to be allocated dynamically.  As the ultimate objective is
not about allowing dodecapus, but about annotating the history upside
down, it also renames "parent" in the code to "scapegoat"; the name of the
game used to be "pass blame to your parents", but now it is "find a
scapegoat to pass blame on".

Signed-off-by: Junio C Hamano <gitster@pobox.com>